refactor: move backfill logic under worker
